+++ tests/settings.py
@@ -32,10 +32,7 @@ INSTALLED_APPS = [
     "django.contrib.messages",
     "django.contrib.staticfiles",
     "paper_uploads",
-    "paper_uploads.cloudinary",
     "django_rq",
-    "django_jinja",
-    "cloudinary",
     "app",
     "examples.fields.standard",
     "examples.fields.custom_storage",
@@ -47,9 +44,6 @@ INSTALLED_APPS = [
     "examples.collections.proxy_models",
     "examples.collections.custom_models",
     "examples.collections.validators",
-    "examples.cloudinary.standard",
-    "examples.cloudinary.custom_storage",
-    "examples.cloudinary.collections",
 ]
 
 MIDDLEWARE = [
@@ -80,10 +74,6 @@ TEMPLATES = [
                 "django.contrib.messages.context_processors.messages",
             ],
         },
-    },
-    {
-        "NAME": "jinja2",
-        "BACKEND": "django_jinja.backend.Jinja2",
     }
 ]
 
@@ -225,29 +215,6 @@ PAPER_MENU = [
                         label=_("Validators"),
                         models=[
                             "validators_collections.Page",
-                        ]
-                    ),
-                ]
-            ),
-            dict(
-                label=_("Cloudinary"),
-                models=[
-                    dict(
-                        label=_("Standard"),
-                        models=[
-                            "standard_cloudinary_fields.Page",
-                        ]
-                    ),
-                    dict(
-                        label=_("Custom destination"),
-                        models=[
-                            "custom_cloudinary_storage.Page",
-                        ]
-                    ),
-                    dict(
-                        label=_("Collections"),
-                        models=[
-                            "cloudinary_collections.Page",
                         ]
                     ),
                 ]